Rewrote main to use Engine Class
Engine class is now responsible for running the main game loop. Main was
rewritten to use the engine to run the game. The ROOT console must be
initialized outside the engine class to avoid segfaults.

Game now runs without crash / segfualting creating a "dungeon" for the
player to move around.