Added prototype room creation to BSPTree
Added a couple functions to fill the BSPTree leaves with rectangular
rooms. However it more often than not segfaults. This is due to a few
reasons:

The split for a given room can be of width 0, which then causes x1 and
x2 (or y1 and y2) to be the same. When the RNG chooses a room
width/height it gets a range where the max is less than the min, causing
a segfault.

Even if the split is not of width 0, if it is too small the above
problem still happens. Need to fix both cases and add better checking to
the room creation.

If no segfaults occur, then maps actually get created. Still a bug in
room placement but at least that doesn't segfault.

The joy of random numbers.