Fixed (Mostly) BSPTree.
I believe most of the issues with the segfaults in the BSPTree have been
fixed. It appears to be functioning correctly with leaf and room
generation, but needs more testing. Creating connections between the
rooms is next.

The Makefile was updated to be fancier. It now just has a list of
objects to build and a better set of rules for debug versus non-debug
output.