BSPTree now at version 1.0
The BSPTree is now at it's first version of features. It still needs
work, especially with the occasionally orphaned rooms it creates, but it
does a decent job of creating random maps.

Besides cleaning / optimizing / rearranging code work needs to be done
on connecting the rooms, new styles of rooms, and probably adding the
room information to the leaves.

The main and engine functions have been edited so that the default map
is now created with the BSPTree. However the character placement is
still fixed and will land the player inside a wall with a decent
probability.

Generating maps with cellular automata has been started, which means
basically empty header and sources files have been created.

The Makefile was edited to include a preprocessor variable for debuging
output. Currently only used in the BSPTree.