Cellular Automata Mostly Finished
The Cellular Automata map generation method is mostly finished. It seems
to be working correctly, although a closer look would be nice. Some more
functionality is planned -- at least making the border a set of walls so
the player can't escape the map and segfault. (Or I could make the
change in the input or map class)

The main functions and engine class were edited to test the CellularMap
generation.

Just to document it but another map generation technique that might be
interesting, especially since I don't think it's been done before, is a
gravity based map generator. The walls or open tiles gravitate toward
each other creating hallways and larger rooms creating a connected map.