Segfaulting even with all new code commented out
Made a first pass at the flood fill algorithm. Upon trying to test the
code crashed with a segfault and now segfaults every time even with all
the new code commented out.

Made this new branch to save the code and to revert back to when it was
working this morning.