Fixed Segfaulting Bug in Map Constructor
Fixed the bug in the Map constructor that was causing a Segfault (Core
Dump) upon execution. The for loop was bounded by vector.size() when the
vector was still at size() = 0. Thus when the grid_to_map() function was
called and tried to delete a tile, there was nothing to delete and
segfault.