Both Cellular and BSP Generation Additions
Implemented a method to create a single cavern room of a minimum area
for the cellular generation. It generates a candidate grid and uses
flood fill to calculate the area of the room. If the area is above
some cut value the grid is accepted, if not, it generates a new grid
and tries again. A parameter can be passed to determine how many
tries are attempted before giving up.

Added the ability to create cave rooms in the leaves of the BSPTree
using the cellular generator.

Added a Room class and functions for later use.