Tried Finding All Rooms in CellularMap
Wrote some code to find all the rooms in a given CellularMap. It seems
to have multiple points where it segfaults, usually resulting in a
double free or something failing with the push_back() function of the
vector.

No idea why it fails here, but works perfectly find in the main()
function. Time to leave that for tomorrow.